:bug: (membership): Fixed customer name not being sent to API Fixed TRI-66
