+++ pages/memberships.js
@@ -5,30 +5,6 @@ import getter from '@/util/getter'
 import groq from 'groq'
 import { Loader } from '@/components/util'
 import PriceCard from '@/components/Membership/PriceCard'
-import axios from 'axios'
-import { notify } from '@/util/notification'
-
-const finishSubmit = (token, user, product) => {
-	const { firstName, lastName, email } = user
-	return axios
-		.post('/api/purchase/membership', {
-			first_name: firstName,
-			last_name: lastName,
-			email,
-			payment_token: token,
-			amount: product.price,
-			sku: product.sku
-		})
-		.then((res) => {
-			console.log(res)
-			notify('success', 'Order successfully placed!')
-			return res.data.payload.transactionID
-		})
-		.catch((err) => {
-			notify('error', err?.message, 'There was an error processing your order')
-			return null
-		})
-}
 
 const MembershipPage = () => {
 	const [product, setProduct] = useState({
@@ -82,80 +58,6 @@ const MembershipPage = () => {
 			</div>
 		</section>
 	)
-}
-
-export const injectCollectJS = (price) => {
-	window.CollectJS.configure({
-		variant: 'inline',
-		paymentType: 'cc',
-		price: price,
-		currency: 'USD',
-		country: 'US',
-		styleSniffer: 'true',
-		fields: {
-			ccnumber: {
-				selector: '#ccnumber',
-				title: 'Card Number',
-				placeholder: '0000 0000 0000 0000'
-			},
-			ccexp: {
-				selector: '#ccexp',
-				title: 'Card Expiration',
-				placeholder: '00 / 00'
-			},
-			cvv: {
-				display: 'show',
-				selector: '#cvv',
-				title: 'CVV Code',
-				placeholder: '***'
-			}
-		},
-		customCss: {
-			border: '0px',
-			'background-color': 'rgb(250 250 250)'
-		},
-		invalidCss: {
-			color: 'red',
-			'background-color': '#f78e83'
-		},
-		validCss: {
-			color: 'green',
-			'background-color': '#d0ffd0'
-		},
-		focusCss: {
-			color: 'black',
-			'background-color': 'rgb(220 220 220)'
-		},
-		placeholderCss: {
-			color: 'black',
-			'background-color': 'rgb(220 220 220)'
-		},
-		validationCallback: function (field, status, message) {
-			if (status) {
-				notify('success', message, field)
-			} else {
-				notify('warning', message, field)
-			}
-		},
-		callback: function (response) {
-			console.log(response)
-			const form = document.getElementsByTagName('form')[0].elements
-			const res = finishSubmit(
-				response.token,
-				{
-					firstName: form['first_name'].value,
-					lastName: form['last_name'].value,
-					email: form['email'].value
-				},
-				{ price: form['price'].value, sku: form['sku'].value }
-			)
-			Promise.resolve(res).then((res) => {
-				if (res) {
-					window.location.href = `/success?transid=${res}`
-				}
-			})
-		}
-	})
 }
 
 export default MembershipPage
